use MIDI music

+++ src/Background.cpp
@@ -2,6 +2,7 @@
 
 #include "Background.h"
 
+#include "defines.h"
 #include "scaler.h"
 #include "Game.h"
 
@@ -10,8 +11,8 @@ Background::Background (const char* filename, int width, int height)
   float stretch_x = g_Game.GetScreen_W () / (float)BASE_SCREEN_X;
   float stretch_y = g_Game.GetScreen_H () / (float)BASE_SCREEN_Y;
 
-  int new_width  = static_cast<int>(width *stretch_x);
-  int new_height = static_cast<int>(height*stretch_y);
+  int new_width  = static_cast<int> (width *stretch_x);
+  int new_height = static_cast<int> (height*stretch_y);
 
   image = LoadScaledBitmap (filename, new_width, new_height);
 
@@ -34,14 +35,14 @@ Background::Background (const char* filename, int width, int height)
 Background::~Background ()
 {
   delete clip;
-  SDL_FreeSurface(image);
+  SDL_FreeSurface (image);
 }
 
 void
 Background::RenderBackground ()
 {
   SDL_Rect offset = {(Sint16)x, (Sint16)y, 0, 0};
-  SDL_BlitSurface(image, clip, g_Game.GetScreen(), &offset);
+  SDL_BlitSurface (image, clip, g_Game.GetScreen (), &offset);
 }
 
 void
@@ -53,6 +54,6 @@ Background::InitBackground ()
 void
 Background::UpdateBackground (float x, float y)
 {
-  Background::x = x - (x * width  / g_GamePtr->GetScreen_W ());
-  Background::y = y - (y * height / g_GamePtr->GetScreen_H ());
+  Background::x = x - (x * width  / static_cast<float> (g_GamePtr->GetScreen_W ()));
+  Background::y = y - (y * height / static_cast<float> (g_GamePtr->GetScreen_H ()));
 }
